Obracanie figury wokół dowolnego punktu

+++ frontend/src/main/java/schumi178/javaprojects/graphics/zad1/shapes/BezierCurve.java
@@ -151,7 +151,7 @@ public class BezierCurve implements DrawableShape {
     }
 
     @Override
-    public void rotate(int angle) {
+    public void rotate(double angle) {
 
     }
