+++ frontend/src/main/java/schumi178/javaprojects/graphics/zad1/shapes/DrawableRectangle.java
@@ -125,7 +125,7 @@ public class DrawableRectangle implements DrawableShape {
     }
 
     @Override
-    public void rotate(int angle) {
+    public void rotate(double angle) {
         Rotate rotate = new Rotate();
         rotate.setAngle(angle);
         rect.setFill(color);
