+++ frontend/src/main/java/schumi178/javaprojects/graphics/zad1/shapes/FreeDrawing.java
@@ -71,7 +71,7 @@ public class FreeDrawing implements DrawableShape {
     }
 
     @Override
-    public void rotate(int angle) {
+    public void rotate(double angle) {
 
     }
